chore: rename comitfuns.go to commit_helper.go for consistency
